fix: rocky problems
